+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * Unreleased
+* 0.9 (2021-08-10)
     * Move Writer classes to
       [AceSegmentWriter](https://github.com/bxparks/AceSegment) library.
 * 0.8.2 (2021-08-09)